Move FederationSupport and PendingFederation outside federation package to avoid circular dependency with peg package
