+++ rskj-core/src/main/java/co/rsk/peg/federation/FederationMember.java
@@ -38,9 +38,52 @@ import java.util.stream.Collectors;
  * @author Ariel Mendelzon
  */
 public final class FederationMember {
+    /**
+     * Compares federation members based on their underlying keys.
+     *
+     * The total ordering is defined such that, for any two members M1, M2,
+     * 1) M1 < M2 iff BTC_PUB_KEY(M1) <lex BTC_PUB_KEY(M2) OR
+     *              (BTC_PUB_KEY(M1) ==lex BTC_PUB_KEY(M2) AND
+     *               RSK_PUB_KEY(M1) <lex RSK_PUB_KEY(M2)) OR
+     *              (BTC_PUB_KEY(M1) ==lex BTC_PUB_KEY(M2) AND
+     *               RSK_PUB_KEY(M1) ==lex RSK_PUB_KEY(M2) AND
+     *               MST_PUB_KEY(M1) <lex MST_PUB_KEY(M2))
+     * 2) M1 == M2 iff BTC_PUB_KEY(M1) ==lex BTC_PUB_KEY(M2) AND
+     *                 RSK_PUB_KEY(M1) ==lex RSK_PUB_KEY(M2) AND
+     *                 MST_PUB_KEY(M1) ==lex MST_PUB_KEY(M2) AND
+     * 3) M1 > M2 otherwise
+     *
+     * where <lex and ==lex is given by negative and zero values (resp.) of the
+     * UnsignedBytes.lexicographicalComparator() comparator.
+     */
+    public static final Comparator<FederationMember> BTC_RSK_MST_PUBKEYS_COMPARATOR = new Comparator<FederationMember>() {
+        private Comparator<byte[]> comparator = UnsignedBytes.lexicographicalComparator();
+
+        @Override
+        public int compare(FederationMember m1, FederationMember m2) {
+            int btcKeysComparison = comparator.compare(m1.getBtcPublicKey().getPubKey(), m2.getBtcPublicKey().getPubKey());
+            if (btcKeysComparison == 0) {
+                int rskKeysComparison = comparator.compare(m1.getRskPublicKey().getPubKey(), m2.getRskPublicKey().getPubKey());
+                if (rskKeysComparison == 0) {
+                    return comparator.compare(m1.getMstPublicKey().getPubKey(), m2.getMstPublicKey().getPubKey());
+                }
+                return rskKeysComparison;
+            }
+            return btcKeysComparison;
+        }
+    };
+
     private final BtcECKey btcPublicKey;
     private final ECKey rskPublicKey;
     private final ECKey mstPublicKey;
+
+    public FederationMember(BtcECKey btcPublicKey, ECKey rskPublicKey, ECKey mstPublicKey) {
+        // Copy public keys to ensure effective immutability
+        // Make sure we always use compressed versions of public keys
+        this.btcPublicKey = BtcECKey.fromPublicOnly(btcPublicKey.getPubKeyPoint().getEncoded(true));
+        this.rskPublicKey = ECKey.fromPublicOnly(rskPublicKey.getPubKey(true));
+        this.mstPublicKey = ECKey.fromPublicOnly(mstPublicKey.getPubKey(true));
+    }
 
     public enum KeyType {
         BTC("btc"),
@@ -81,49 +124,6 @@ public final class FederationMember {
 
     public static List<FederationMember> getFederationMembersFromKeys(List<BtcECKey> pks) {
         return pks.stream().map(pk -> getFederationMemberFromKey(pk)).collect(Collectors.toList());
-    }
-
-    /**
-     * Compares federation members based on their underlying keys.
-     *
-     * The total ordering is defined such that, for any two members M1, M2,
-     * 1) M1 < M2 iff BTC_PUB_KEY(M1) <lex BTC_PUB_KEY(M2) OR
-     *              (BTC_PUB_KEY(M1) ==lex BTC_PUB_KEY(M2) AND
-     *               RSK_PUB_KEY(M1) <lex RSK_PUB_KEY(M2)) OR
-     *              (BTC_PUB_KEY(M1) ==lex BTC_PUB_KEY(M2) AND
-     *               RSK_PUB_KEY(M1) ==lex RSK_PUB_KEY(M2) AND
-     *               MST_PUB_KEY(M1) <lex MST_PUB_KEY(M2))
-     * 2) M1 == M2 iff BTC_PUB_KEY(M1) ==lex BTC_PUB_KEY(M2) AND
-     *                 RSK_PUB_KEY(M1) ==lex RSK_PUB_KEY(M2) AND
-     *                 MST_PUB_KEY(M1) ==lex MST_PUB_KEY(M2) AND
-     * 3) M1 > M2 otherwise
-     *
-     * where <lex and ==lex is given by negative and zero values (resp.) of the
-     * UnsignedBytes.lexicographicalComparator() comparator.
-     */
-    public static final Comparator<FederationMember> BTC_RSK_MST_PUBKEYS_COMPARATOR = new Comparator<FederationMember>() {
-        private Comparator<byte[]> comparator = UnsignedBytes.lexicographicalComparator();
-
-        @Override
-        public int compare(FederationMember m1, FederationMember m2) {
-            int btcKeysComparison = comparator.compare(m1.getBtcPublicKey().getPubKey(), m2.getBtcPublicKey().getPubKey());
-            if (btcKeysComparison == 0) {
-                int rskKeysComparison = comparator.compare(m1.getRskPublicKey().getPubKey(), m2.getRskPublicKey().getPubKey());
-                if (rskKeysComparison == 0) {
-                    return comparator.compare(m1.getMstPublicKey().getPubKey(), m2.getMstPublicKey().getPubKey());
-                }
-                return rskKeysComparison;
-            }
-            return btcKeysComparison;
-        }
-    };
-
-    public FederationMember(BtcECKey btcPublicKey, ECKey rskPublicKey, ECKey mstPublicKey) {
-        // Copy public keys to ensure effective immutability
-        // Make sure we always use compressed versions of public keys
-        this.btcPublicKey = BtcECKey.fromPublicOnly(btcPublicKey.getPubKeyPoint().getEncoded(true));
-        this.rskPublicKey = ECKey.fromPublicOnly(rskPublicKey.getPubKey(true));
-        this.mstPublicKey = ECKey.fromPublicOnly(mstPublicKey.getPubKey(true));
     }
 
     public BtcECKey getBtcPublicKey() {
