+++ rskj-core/src/test/java/co/rsk/peg/BridgeSupportAddSignatureTest.java
@@ -7,7 +7,6 @@ import java.util.*;
 import co.rsk.config.BridgeConstants;
 import co.rsk.peg.federation.Federation;
 import co.rsk.peg.federation.FederationFactory;
-import co.rsk.peg.federation.FederationSupport;
 import co.rsk.peg.federation.FederationTestUtils;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
 import org.ethereum.config.blockchain.upgrades.ConsensusRule;
