+++ rskj-core/src/test/java/co/rsk/peg/federation/NonStandardErpFederationsTest.java
@@ -139,6 +139,31 @@ class NonStandardErpFederationsTest {
     }
 
     @ParameterizedTest
+    @ValueSource(ints = {-100, 0})
+    void createFederation_withInvalidThresholdValues_throwsIllegalArgumentException(int threshold) {
+        ErpRedeemScriptBuilder builder = new NonStandardErpRedeemScriptBuilder();
+        assertThrows(IllegalArgumentException.class,
+            () -> builder.createRedeemScriptFromKeys(
+                defaultKeys, threshold,
+                emergencyKeys, emergencyThreshold,
+                activationDelayValue)
+        );
+    }
+
+    @Test
+    void createFederation_withThresholdAboveDefaultKeysSize_throwsIllegalArgumentException() {
+        int defaultThresholdAboveDefaultKeysSize = defaultKeys.size() + 1;
+
+        ErpRedeemScriptBuilder builder = new NonStandardErpRedeemScriptBuilder();
+        assertThrows(IllegalArgumentException.class,
+            () -> builder.createRedeemScriptFromKeys(
+                defaultKeys, defaultThresholdAboveDefaultKeysSize,
+                emergencyKeys, emergencyThreshold,
+                activationDelayValue)
+        );
+    }
+
+    @ParameterizedTest
     @ValueSource(longs = { 130L, 500L, 33_000L, ErpRedeemScriptBuilderUtils.MAX_CSV_VALUE})
     void createFederation_postRSKIP293_withValidCsvValues_valid(long csvValue) {
         when(activations.isActive(ConsensusRule.RSKIP284)).thenReturn(true);
