+++ rskj-core/src/test/java/co/rsk/peg/performance/FederationChangeTest.java
@@ -23,7 +23,7 @@ import co.rsk.bitcoinj.store.BtcBlockStore;
 import co.rsk.core.RskAddress;
 import co.rsk.peg.*;
 import co.rsk.peg.federation.FederationMember;
-import co.rsk.peg.federation.PendingFederation;
+import co.rsk.peg.PendingFederation;
 import org.ethereum.core.Repository;
 import org.ethereum.crypto.ECKey;
 import org.ethereum.crypto.HashUtil;
